+++ Xext/render/animcur.c
