+++ Xext/render/glyphstr_priv.h
