+++ Xext/render/picture.c
